requirements.txt
mysql-connector-python
better-profanity